+++ internal/api/k8s/cluster.go
@@ -19,27 +19,27 @@ func ParseComponentImage(kubectlExecOutput string, imageSection string) (string,
 }
 
 // TODO add spec for this
-func KubectlGetImageCommand(k8sObject string, component string) string {
+func KubectlGetImageCommand(k8sObject, component, namespace string) string {
 	return fmt.Sprintf(`
 	kubectl
 	get
 	%s
 	%s
-	--namespace kube-system
+	--namespace %s
 	-o=jsonpath='{$.spec.template.spec.containers[:1].image}'
-	`, k8sObject, component)
+	`, k8sObject, component, namespace)
 }
 
 // TODO add spec for this
-func KubectlSetImageCommand(k8sObject string, componentName string, containerImage string) string {
+func KubectlSetImageCommand(k8sObject, componentName, containerImage, namespace string) string {
 	return fmt.Sprintf(`
 	kubectl
 	set
 	image
 	%s
-	--namespace kube-system
+	--namespace %s
 	%s=%s
-	`, k8sObject, componentName, containerImage)
+	`, k8sObject, componentName, containerImage, namespace)
 }
 
 // TODO add spec for this