+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - the config keys being used, please refer the same config and rename the key attributes.
 
 #### Adds
-- ability to read container image to be read from config for a component being updated, for cases when deployment name
+- ability to read container image, namespace to be read from config for a component being updated, for cases when deployment name
 is not the same as the container name for the container for which update is being made.
 
 ### v0.1.0